Started settings menu code refactoring
