+++ Celestial_Corruption/Assets/Scripts/UI_Scripts/CamSensitivityMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87b732457050f2b61b622cd4e864a690
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
